+++ lib/reflection_factory.dart
@@ -2,7 +2,9 @@
 library reflection_factory;
 
 export 'dart:async';
+
 export 'package:pub_semver/pub_semver.dart';
 
 export 'src/reflection_factory_annotation.dart';
 export 'src/reflection_factory_base.dart';
+export 'src/reflection_factory_extension.dart';