+++ test/src/user_simple.dart
@@ -2,6 +2,8 @@ class TestUserSimple {
   static final double version = 1.0;
   static final bool withReflection = false;
 
+  static bool isVersion(double ver) => version == ver;
+
   final String name;
 
   String? email;
@@ -10,5 +12,11 @@ class TestUserSimple {
 
   TestUserSimple(this.name, this.email, this.password);
 
-  bool checkThePassword(String password) => this.password == password;
+  bool checkThePassword(String password, {bool ignoreCase = false}) {
+    if (ignoreCase) {
+      return this.password.toLowerCase() == password.toLowerCase();
+    } else {
+      return this.password == password;
+    }
+  }
 }