+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 1.0.2
+
+- `ClassReflection`:
+  - Added: `allMethods`, `allStaticMethods`
+- `MethodReflection`:
+  - Methods parameters now are defined with `ParameterReflection`.
+  - `returnType`: fixed for `void`.
+  - `method`: exposed, not private anymore.
+  - Added:
+    - `equalsNormalParametersTypes`, `equalsOptionalParametersTypes`, `equalsNamedParametersTypes`.
+    - `methodInvocation`, `methodInvocationFromMap`.
+- `MethodInvocation`: class to represent an invocation. 
+- Builder:
+  - Now generates documentation of generated extension methods.
+  - Better handling of parameters and fields with `ParameterReflection`.
+    - Now knows if a parameter is `required`. 
+- Improved API documentation.
+- Improved tests.
+
 ## 1.0.1
 
 - @EnableReflection: